// naomi-seeds-backend/src/modules/roles/roles.enum.ts
export enum Role {
    ADMIN = "ADMIN",
    DIRECTOR = "DIRECTOR",
    OPERATIONAL_MANAGER = "OPERATIONAL_MANAGER",
    SALES_OFFICER = "SALES_OFFICER",
    DISTRIBUTOR = "DISTRIBUTOR",
    MANAGER = "MANAGER",
    ACCOUNTANT = "ACCOUNTANT",
    PLANT_MANAGER = "PLANT_MANAGER",
    COMPANY = "COMPANY"
  }



  export enum IRoleMessage {
    unAuthorisedAccess = "Unauthorised Access",
}