// naomi-seeds-backend/src/modules/users/user-role.enum.ts
export enum RoleUpdated {
    ADMIN = "ADMIN",
    DIRECTOR = "DIRECTOR",
    OPERATIONAL_MANAGER = "OPERATIONAL_MANAGER",
    SALES_OFFICER = "SALES_OFFICER",
    DISTRIBUTOR = "DISTRIBUTOR",
    MANAGER = "MANAGER",
    ACCOUNTANT = "ACCOUNTANT",
    PLANT_MANAGER = "PLANT_MANAGER",
  }



  export enum IRoleMessage {
    unAuthorisedAccess = "Unauthorised Access",
}